Revert automated_test_runner.py to original version and enable LLM-based evaluation by default in add_llm_evaluation.py
